Add entity spawning to YAML.

+++ BasicRPGTest_Mono/Engine/Entities/Spawn.cs
@@ -7,8 +7,8 @@ namespace BasicRPGTest_Mono.Engine.Entities
     public class Spawn
     {
         public LivingEntity entity;
-        public int weight;
-        public Spawn(LivingEntity entity, int weight)
+        public double weight;
+        public Spawn(LivingEntity entity, double weight)
         {
             this.entity = entity;
             this.weight = weight;
